# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AgroAI Smart Greenhouse API System",
    "description": "Backend API for AgroAI Smart Greenhouse API System",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/greenhouses/{greenhouse_id}/devices": {
      "get": {
        "tags": [
          "greenhouses",
          "devices"
        ],
        "summary": "List Devices",
        "operationId": "list_devices_api_greenhouses__greenhouse_id__devices_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DeviceRead"
                  },
                  "title": "Response List Devices Api Greenhouses  Greenhouse Id  Devices Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/greenhouses/{greenhouse_id}/devices/{device_name}/switch/{device_state}": {
      "post": {
        "tags": [
          "greenhouses",
          "devices"
        ],
        "summary": "Device Switch On Off",
        "operationId": "device_switch_on_off_api_greenhouses__greenhouse_id__devices__device_name__switch__device_state__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_name",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/SwitchableDeviceName"
            }
          },
          {
            "name": "device_state",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "off",
                "on"
              ],
              "type": "string",
              "title": "Device State"
            }
          },
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/greenhouses/{greenhouse_id}/devices/{device_name}/settings": {
      "post": {
        "tags": [
          "greenhouses",
          "devices"
        ],
        "summary": "Device Settings",
        "description": "Update device settings.",
        "operationId": "device_settings_api_greenhouses__greenhouse_id__devices__device_name__settings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_name",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ConfigurableDeviceName"
            }
          },
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceSettingsModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/greenhouses/{greenhouse_id}/devices/settings": {
      "post": {
        "tags": [
          "greenhouses",
          "devices"
        ],
        "summary": "Bulk Device Settings",
        "description": "Update multiple sensor thresholds as one settings operation.",
        "operationId": "bulk_device_settings_api_greenhouses__greenhouse_id__devices_settings_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "greenhouse_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Greenhouse Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeviceSettingsModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BulkDeviceSettingsModel": {
        "properties": {
          "air": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceSettingsModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "humidity": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceSettingsModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "temperature": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceSettingsModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "moisture": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceSettingsModel"
              },
              {
                "type": "null"
              }
            ]
          },
          "light": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeviceSettingsModel"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "BulkDeviceSettingsModel"
      },
...
      "CommandResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "command_id": {
            "type": "string",
            "title": "Command Id"
          },
          "status": {
            "$ref": "#/components/schemas/CommandStatus"
          }
        },
        "type": "object",
        "required": [
          "command_id",
          "status"
        ],
        "title": "CommandResponse"
      },
      "CommandStatus": {
        "type": "string",
        "enum": [
          "pending",
          "published",
          "acknowledged",
          "failed",
          "timed_out"
        ],
        "title": "CommandStatus"
      },
      "ConfigurableDeviceName": {
        "type": "string",
        "enum": [
          "air",
          "humidity",
          "temperature",
          "moisture",
          "light"
        ],
        "title": "ConfigurableDeviceName"
      },
      "DeviceRead": {
        "properties": {
          "greenhouse_id": {
            "type": "integer",
            "title": "Greenhouse Id"
          },
          "type": {
            "$ref": "#/components/schemas/DeviceType"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "topic_root": {
            "type": "string",
            "title": "Topic Root"
          },
          "min_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Value"
          },
          "max_value": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Value"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "greenhouse_id",
          "type",
          "name",
          "topic_root",
          "id"
        ],
        "title": "DeviceRead"
      },
      "DeviceSettingsModel": {
        "properties": {
          "min": {
            "type": "number",
            "title": "Min"
          },
          "max": {
            "type": "number",
            "title": "Max"
          }
        },
        "type": "object",
        "required": [
          "min",
          "max"
        ],
        "title": "DeviceSettingsModel"
      },
      "DeviceType": {
        "type": "string",
        "enum": [
          "SENSOR",
          "ACTUATOR"
        ],
        "title": "DeviceType"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SwitchableDeviceName": {
        "type": "string",
        "enum": [
          "soil_water_pump",
          "air_water_pump",
          "led",
          "fan"
        ],
        "title": "SwitchableDeviceName"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}